Change initCorners for MidpointDisplacement
Preseed two randomly selected corners with values taken from the app.config.

This is to ensure we do not have all land or all water maps and also to allow through configuration to have more control over the resulting maps as the initial corner values are determinant in all the final map will look like.